Cleaning up the code and ensuring other weights wont crash
