+++ atom/quantization/quark/utils.py
@@ -105,98 +105,3 @@ def _is_equal_or_regex_match(value: str,
     elif target == value:
         return True
     return False
-
-# utility for tensor dims > 2 cases
-def b_dynamic_mxfp4_quant(x):
-    h, b, d = x.shape
-    x, x_scales = dynamic_mxfp4_quant(x.reshape(-1, d))
-    return x.view(h, b, d // 2), x_scales.view(h, b, d // 32)
-    #return x.view(h, b, d // 2), x_scales.view(h, b, d // 32)
-
-
-def mxfp4_to_f32(x, is_threed):
-    # 2 because we pack fp4 in uint8.
-    x = x.repeat_interleave(2, dim=-1)
-    if is_threed:
-        x[..., ::2] = x[..., ::2] & 0xF
-        x[..., 1::2] = x[..., 1::2] >> 4
-    else:
-        x[:, ::2] = x[:, ::2] & 0xF
-        x[:, 1::2] = x[:, 1::2] >> 4
-
-    mxfp4_list = [
-        0.0,
-        0.5,
-        1.0,
-        1.5,
-        2.0,
-        3.0,
-        4.0,
-        6.0,
-        -0.0,
-        -0.5,
-        -1.0,
-        -1.5,
-        -2.0,
-        -3.0,
-        -4.0,
-        -6.0,
-    ]
-    mxfp4_in_f32 = torch.tensor(mxfp4_list, dtype=torch.float32, device="cuda")
-    return mxfp4_in_f32[x.long()]
-
-
-def e8m0_to_f32(x):
-    # Convert the input tensor `x` (assumed to be in e8m0 format) to float32.
-    # e8m0 is a custom 8-bit floating point format with 8 bits for exponent, 0 for mantissa.
-    # This means the value is essentially 2^(exponent - 127), similar to how IEEE-754 stores floats.
-
-    # Convert x to float32 for computation, and compute the power of 2 by subtracting the bias (127).
-    x_f32 = 2 ** ((x.to(torch.float32)) - 127)
-
-    # If the exponent value was 255 (i.e., 2^(128)), this is a special case usually used to represent NaN or Inf.
-    # Since this custom format has no mantissa, treat 2^128 as NaN.
-    x_f32[x_f32 == 128] = float("nan")
-    return x_f32
-
-def quark_post_load_weights(self_attn: nn.Module, w: torch.Tensor, quant_format: str):
-    if "mxfp4" in quant_format:
-     
-        # when dtype is bf16, the processing flow is to dynamic quantize bf16 tensor to uint8 tensor
-        # do w_kc (bf16) first to get the w_kc(uint8) w_s_kc(uint8)
-        # and w_vc repeating the same procedure of w_kc to get  w_vc(uint8) w_s_vc(uint8)
-        if w.dtype == torch.bfloat16:
-            # w_kc, w_vc = w.split(
-            # [self_attn.qk_nope_head_dim, self_attn.v_head_dim], dim=1)
-            w_kc, w_vc = w.unflatten(
-                0, (-1, self_attn.qk_nope_head_dim + self_attn.v_head_dim)
-            ).split([self_attn.qk_nope_head_dim, self_attn.v_head_dim], dim=1)
-            w_kc, w_s_kc = b_dynamic_mxfp4_quant(w_kc.transpose(-2, -1))
-            w_kc = w_kc.transpose(-2, -1)
-            w_s_kc = w_s_kc.transpose(-2, -1)
-            w_vc, w_s_vc = b_dynamic_mxfp4_quant(w_vc)
-            w_s_kc = w_s_kc.transpose(1, 2).contiguous().transpose(1, 2)
-            w_s_vc = w_s_vc.contiguous().transpose(1, 2)
-        elif w.dtype == torch.uint8:  # static quant for mxfp4
-            # when dtype is uint8, it means the w has been quantized to mxfp4 format
-            # but we must separate it to w_kc and w_vc.
-            # The quantized tensor size is only half of original tensor size
-            # and the scaling factor is 1/32, the transpose behavior will be not correct
-            # need to upcast it to fp32 to separate w to w_kc and w_vc
-            # to ensure the following transpose behavior is correct
-            # and then do mxfp4 quant again
-            w = mxfp4_to_f32(w, True).to(torch.bfloat16)
-            w_scales = self_attn.kv_b_proj.weight_scale.repeat_interleave(32, dim=-1)
-            w_scales = e8m0_to_f32(w_scales).to(torch.bfloat16)
-            w = w * w_scales
-            w_kc, w_vc = w.unflatten(
-                0, (-1, (self_attn.qk_nope_head_dim + self_attn.v_head_dim))
-            ).split([self_attn.qk_nope_head_dim, self_attn.v_head_dim], dim=1)
-            w_kc, w_s_kc = b_dynamic_mxfp4_quant(w_kc.transpose(-2, -1))
-            w_kc = w_kc.transpose(-2, -1)
-            w_s_kc = w_s_kc.transpose(-2, -1)
-            w_vc, w_s_vc = b_dynamic_mxfp4_quant(w_vc)
-            w_s_kc = w_s_kc.transpose(1, 2).contiguous().transpose(1, 2)
-            w_s_vc = w_s_vc.contiguous().transpose(1, 2)
-
-        return w_kc, w_s_kc, w_vc, w_s_vc
